Programa definitivo

+++ es.ull.etsii/src/vision_por_computador/Matriz.java
@@ -2,14 +2,14 @@ package vision_por_computador;
 
 public class Matriz {
   
-  private double[][] A;
+  private float[][] A;
   private int m;
   private int p;
   
-  public Matriz(double[][] B) {
+  public Matriz(float[][] B) {
     this.m = B.length;
     this.p = B[0].length;
-    this.A = new double[this.m][this.p];
+    this.A = new float[this.m][this.p];
     for (int i = 0; i < this.m; i++) {
       for (int j = 0; j < this.p; j++) {
         this.A[i][j] = B[i][j];
@@ -18,9 +18,9 @@ public class Matriz {
   }
   
   public Matriz producto(Matriz B) {
-    double[][] c = new double[this.m][B.getP()];
-    double[][] a = this.getMatriz();
-    double[][] b = B.getMatriz();
+    float[][] c = new float[this.m][B.getP()];
+    float[][] a = this.getMatriz();
+    float[][] b = B.getMatriz();
     
     for (int i = 0; i < this.m; i++) {
       for (int j = 0; j < c[0].length; j++) {
@@ -30,8 +30,8 @@ public class Matriz {
     return (new Matriz(c));
   }
   
-  private double sumatorio(double[][] a, double[][] b, int i, int j) {
-    double result = 0;
+  private float sumatorio(float[][] a, float[][] b, int i, int j) {
+    float result = 0;
     for (int k = 0; k < b.length; k++) {
       result += a[i][k] * b[k][j];
     }
@@ -58,7 +58,7 @@ public class Matriz {
     return (this.p);
   }
   
-  public double[][] getMatriz() {
+  public float[][] getMatriz() {
     return (this.A);
   }
